move the entity subscriber to db dir
